Refactors code by moving the models from individual pages to shared folder
